+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * Raise `CompilerError` errors when no clauses nor blocks are provided.
   * Export formatter configuration via `.formatter.exs`.
   * Support single line usage (i.e. `async with a <- 1, do: a`).
+  * Re-throw uncaught values (i.e. `async with _ <- throw(:foo), do: :ok`).
 
 ## v0.2.2